separate backend files into API and DB
